Moved modeling into individual files
Removed account tests

Removed appsettings for localhost url
