+++ src/API/LeadershipProfileAPI/Data/Models/LogoutInputModel.cs
@@ -1,0 +1,7 @@
+﻿namespace LeadershipProfileAPI.Data.Models
+{
+    public class LogoutInputModel
+    {
+        public string LogoutId { get; set; }
+    }
+}
