+++ src/API/LeadershipProfileAPI/appsettings.json
@@ -11,8 +11,5 @@
     "Sender": "\"Leadership Portal\" <noreply@test.com>",
     "Username": "",
     "Password": ""
-  },
-  "ApplicationConfiguration": {
-    "WebApplicationUrl": "https://localhost:5001"
   }
 }
